+++ rinex-cli/src/cli/qc.rs
@@ -1,6 +1,6 @@
 // tbin opmode
-use clap::{value_parser, Arg, ArgAction, ArgMatches, ColorChoice, Command};
-use rinex::prelude::Duration;
+use clap::{Arg, ArgAction, Command};
+
 
 pub fn subcommand() -> Command {
     Command::new("qc")
@@ -11,11 +11,20 @@ pub fn subcommand() -> Command {
 This is typically used prior precise point positioning.",
         )
         .arg(
+            Arg::new("spp")
+                .long("spp")
+                .action(ArgAction::SetTrue)
+                .help("Force solving method to SPP.
+Otherwise we use the default Method.
+See online documentations [https://docs.rs/gnss-rtk/latest/gnss_rtk/prelude/enum.Method.html#variants]."))
+        .arg(
             Arg::new("cfg")
+                .short('c')
                 .long("cfg")
                 .required(false)
                 .value_name("FILE")
                 .action(ArgAction::Append)
-                .help("Pass a QC configuration file (JSON)."),
-        )
+                .help("Pass a QC configuration file (JSON).
+[] is the structure to represent in JSON.
+See [] for meaningful examples."))
 }
