+++ rinex-cli/src/cli/split.rs
@@ -1,7 +1,7 @@
 // Merge opmode
-use clap::{value_parser, Arg, ArgAction, ArgMatches, ColorChoice, Command};
+use clap::{value_parser, Arg, ArgAction, Command};
 use rinex::prelude::Epoch;
-use std::path::PathBuf;
+
 
 pub fn subcommand() -> Command {
     Command::new("split")
