+++ rinex-cli/src/graph/mod.rs
@@ -23,7 +23,7 @@ use plotly::{
 use rand::Rng;
 use serde::Serialize;
 
-use rinex::prelude::RnxContext;
+
 use rinex::prelude::*;
 
 mod record;
@@ -36,7 +36,7 @@ mod skyplot;
 use skyplot::skyplot;
 
 mod naviplot;
-use naviplot::naviplot;
+
 
 mod combination;
 use combination::{plot_gnss_code_mp, plot_gnss_combination, plot_gnss_dcb};
