+++ rinex-cli/src/positioning/mod.rs
@@ -1,5 +1,261 @@
-pub mod solver;
-pub use solver::solver;
+use crate::cli::Context;
+use std::fs::{read_to_string, File};
+//use statrs::statistics::Statistics;
 
-pub mod post_process;
-pub use post_process::post_process;
+mod ppp; // precise point positioning
+
+use gnss::prelude::Constellation; // SV};
+use rinex::carrier::Carrier;
+use rinex::navigation::Ephemeris;
+use rinex::prelude::{Observable, Rinex, RnxContext};
+use clap::ArgMatches;
+
+use rtk::prelude::{
+    AprioriPosition, BdModel, Candidate, Config, Duration, Epoch, InterpolationResult,
+    IonosphericBias, KbModel, Method, NgModel, Observation, PVTSolution, PVTSolutionType, Solver,
+    TroposphericBias, Vector3,
+};
+
+use map_3d::{ecef2geodetic, Ellipsoid};
+use std::collections::{BTreeMap, HashMap};
+use thiserror::Error;
+
+#[derive(Debug, Error)]
+pub enum Error {
+    #[error("solver error")]
+    SolverError(#[from] rtk::Error),
+    #[error("undefined apriori position")]
+    UndefinedAprioriPosition,
+}
+
+fn tropo_components(meteo: Option<&Rinex>, t: Epoch, lat_ddeg: f64) -> Option<(f64, f64)> {
+    const MAX_LATDDEG_DELTA: f64 = 15.0;
+    let max_dt = Duration::from_hours(24.0);
+    let rnx = meteo?;
+    let meteo = rnx.header.meteo.as_ref().unwrap();
+
+    let delays: Vec<(Observable, f64)> = meteo
+        .sensors
+        .iter()
+        .filter_map(|s| match s.observable {
+            Observable::ZenithDryDelay => {
+                let (x, y, z, _) = s.position?;
+                let (lat, _, _) = ecef2geodetic(x, y, z, Ellipsoid::WGS84);
+                if (lat - lat_ddeg).abs() < MAX_LATDDEG_DELTA {
+                    let value = rnx
+                        .zenith_dry_delay()
+                        .filter(|(t_sens, _)| (*t_sens - t).abs() < max_dt)
+                        .min_by_key(|(t_sens, _)| (*t_sens - t).abs());
+                    let (_, value) = value?;
+                    debug!("{:?} lat={} zdd {}", t, lat_ddeg, value);
+                    Some((s.observable.clone(), value))
+                } else {
+                    None
+                }
+            },
+            Observable::ZenithWetDelay => {
+                let (x, y, z, _) = s.position?;
+                let (lat, _, _) = ecef2geodetic(x, y, z, Ellipsoid::WGS84);
+                if (lat - lat_ddeg).abs() < MAX_LATDDEG_DELTA {
+                    let value = rnx
+                        .zenith_wet_delay()
+                        .filter(|(t_sens, _)| (*t_sens - t).abs() < max_dt)
+                        .min_by_key(|(t_sens, _)| (*t_sens - t).abs());
+                    let (_, value) = value?;
+                    debug!("{:?} lat={} zdd {}", t, lat_ddeg, value);
+                    Some((s.observable.clone(), value))
+                } else {
+                    None
+                }
+            },
+            _ => None,
+        })
+        .collect();
+
+    if delays.len() < 2 {
+        None
+    } else {
+        let zdd = delays
+            .iter()
+            .filter_map(|(obs, value)| {
+                if obs == &Observable::ZenithDryDelay {
+                    Some(*value)
+                } else {
+                    None
+                }
+            })
+            .reduce(|k, _| k)
+            .unwrap();
+
+        let zwd = delays
+            .iter()
+            .filter_map(|(obs, value)| {
+                if obs == &Observable::ZenithWetDelay {
+                    Some(*value)
+                } else {
+                    None
+                }
+            })
+            .reduce(|k, _| k)
+            .unwrap();
+
+        Some((zwd, zdd))
+    }
+}
+
+fn kb_model(nav: &Rinex, t: Epoch) -> Option<KbModel> {
+    let kb_model = nav
+        .klobuchar_models()
+        .min_by_key(|(t_i, _, _)| (t - *t_i).abs());
+
+    if let Some((_, sv, kb_model)) = kb_model {
+        Some(KbModel {
+            h_km: {
+                match sv.constellation {
+                    Constellation::BeiDou => 375.0,
+                    // we only expect GPS or BDS here,
+                    // badly formed RINEX will generate errors in the solutions
+                    _ => 350.0,
+                }
+            },
+            alpha: kb_model.alpha,
+            beta: kb_model.beta,
+        })
+    } else {
+        /* RINEX 3 case */
+        let iono_corr = nav.header.ionod_correction?;
+        if let Some(kb_model) = iono_corr.as_klobuchar() {
+            Some(KbModel {
+                h_km: 350.0, //TODO improve this
+                alpha: kb_model.alpha,
+                beta: kb_model.beta,
+            })
+        } else {
+            None
+        }
+    }
+}
+
+fn bd_model(nav: &Rinex, t: Epoch) -> Option<BdModel> {
+    nav.bdgim_models()
+        .min_by_key(|(t_i, _)| (t - *t_i).abs())
+        .map(|(_, model)| BdModel { alpha: model.alpha })
+}
+
+fn ng_model(nav: &Rinex, t: Epoch) -> Option<NgModel> {
+    nav.nequick_g_models()
+        .min_by_key(|(t_i, _)| (t - *t_i).abs())
+        .map(|(_, model)| NgModel { a: model.a })
+}
+
+pub fn precise_positioning(ctx: &Context, matches: &ArgMatches) -> Result<(), Error> {
+    
+    let method = match matches.get_flag("spp") {
+        true => Method::SPP,
+        false => Method::PPP,
+    };
+
+    let cfg = match matches.get_one::<String>("cfg") {
+        Some(fp) => {
+            let content = read_to_string(fp)
+                .unwrap_or_else(|_| panic!("failed to read configuration: permission denied"));
+            let cfg = serde_json::from_str(&content)
+                .unwrap_or_else(|_| panic!("failed to parse configuration: invalid content"));
+            info!("using custom solver configuration: {:#?}", cfg);
+            cfg
+        },
+        None => {
+            let cfg = Config::preset(method);
+            info!("using default {:?} solver preset: {:#?}", method, cfg);
+            cfg
+        },
+    };
+
+    let apriori_ecef = ctx.rx_ecef
+        .ok_or(Error::UndefinedAprioriPosition)?;
+
+    let apriori = Vector3::<f64>::new(apriori_ecef.0, apriori_ecef.1, apriori_ecef.2);
+    let apriori = AprioriPosition::from_ecef(apriori);
+    let lat_ddeg = apriori.geodetic[0];
+
+    /*
+     * verify requirements
+     */
+    if ctx.data.obs_data().is_none() {
+        panic!("positioning requires Observation RINEX");
+    }
+
+    let nav_data = ctx.data.nav_data()
+        .expect("positioning requires Navigation RINEX");
+
+    let meteo_data = ctx.data.meteo_data();
+
+    let sp3_data = ctx.data.sp3_data()
+    let sp3_has_clock = match sp3_data {
+        Some(sp3) => sp3.sv_clock().count() > 0,
+        None => false,
+    };
+        
+    if sp3_data.is_none() {
+        panic!(
+        "High precision orbits (SP3) are unfortunately mandatory at the moment..");
+    };
+
+    // print config to be used
+    info!("Using solver {:?} method", method);
+    info!("Using solver configuration {:#?}", cfg);
+
+    let mut solver = Solver::new(
+        &cfg,
+        apriori,
+        /* state vector interpolator */
+        |t, sv, order| {
+            /* SP3 source is prefered */
+            if let Some(sp3) = sp3_data {
+                if let Some((x, y, z)) = sp3.sv_position_interpolate(sv, t, order) {
+                    let (x, y, z) = (x * 1.0E3, y * 1.0E3, z * 1.0E3);
+                    let (elevation, azimuth) =
+                        Ephemeris::elevation_azimuth((x, y, z), apriori_ecef);
+                    Some(
+                        InterpolationResult::from_mass_center_position((x, y, z))
+                            .with_elevation_azimuth((elevation, azimuth)),
+                    )
+                } else {
+                    // debug!("{:?} ({}): sp3 interpolation failed", t, sv);
+                    if let Some((x, y, z)) = nav_data.sv_position_interpolate(sv, t, order) {
+                        let (x, y, z) = (x * 1.0E3, y * 1.0E3, z * 1.0E3);
+                        let (elevation, azimuth) =
+                            Ephemeris::elevation_azimuth((x, y, z), apriori_ecef);
+                        Some(
+                            InterpolationResult::from_apc_position((x, y, z))
+                                .with_elevation_azimuth((elevation, azimuth)),
+                        )
+                    } else {
+                        // debug!("{:?} ({}): nav interpolation failed", t, sv);
+                        None
+                    }
+                }
+            } else if let Some((x, y, z)) = nav_data.sv_position_interpolate(sv, t, order) {
+                let (x, y, z) = (x * 1.0E3, y * 1.0E3, z * 1.0E3);
+                let (elevation, azimuth) = Ephemeris::elevation_azimuth((x, y, z), apriori_ecef);
+                Some(
+                    InterpolationResult::from_apc_position((x, y, z))
+                        .with_elevation_azimuth((elevation, azimuth)),
+                )
+            } else {
+                // debug!("{:?} ({}): nav interpolation failed", t, sv);
+                None
+            }
+        },
+        /* APC corrections provider */
+        |_t, _sv, _freq| None,
+    )?;
+
+    if matches.get_flag("cggtts") {
+        /* CGGTTS special opmode */
+    } else {
+        /* PPP */
+        let pvt_solutions = ppp::resolve(&ctx, solver);
+    }
+    Ok(())
+}
